requirements.txt
streamlit
python-dotenv
requests
newsapi-python
tavily-python
langchain
langgraph
langchain-openai
langchain-community
